backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
soundfile>=0.12.0
librosa>=0.10.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pydantic>=2.0.0
hdbscan>=0.8.33
umap-learn>=0.5.3
